requirements.txt
streamlit
numpy
pandas
plotly
mysql-connector-python
altair
python-dotenv
